+++ tests/tests/regression/issue_3457.rs
@@ -144,6 +144,8 @@ fn pass_reset_vertex_buffer() {
                 },
             })],
             depth_stencil_attachment: None,
+            timestamp_writes: None,
+            occlusion_query_set: None,
         });
 
         double_rpass.set_pipeline(&double_pipeline);
@@ -177,6 +179,8 @@ fn pass_reset_vertex_buffer() {
                 },
             })],
             depth_stencil_attachment: None,
+            timestamp_writes: None,
+            occlusion_query_set: None,
         });
 
         single_rpass.set_pipeline(&single_pipeline);