+++ wgpu-core/src/command/clear.rs
@@ -452,6 +452,8 @@ fn clear_texture_via_render_passes<A: hal::Api>(
                     color_attachments,
                     depth_stencil_attachment,
                     multiview: None,
+                    timestamp_writes: None,
+                    occlusion_query_set: None,
                 });
                 encoder.end_render_pass();
             }