+++ wgpu-core/src/command/draw.rs
@@ -243,6 +243,10 @@ pub enum RenderCommand {
         query_set_id: id::QuerySetId,
         query_index: u32,
     },
+    BeginOcclusionQuery {
+        query_index: u32,
+    },
+    EndOcclusionQuery,
     BeginPipelineStatisticsQuery {
         query_set_id: id::QuerySetId,
         query_index: u32,