+++ wgpu-hal/src/vulkan/mod.rs
@@ -96,6 +96,10 @@ pub struct DebugUtilsMessengerUserData {
 
     /// Validation layer specification version, from `vk::LayerProperties`.
     validation_layer_spec_version: u32,
+
+    /// If the OBS layer is present. OBS never increments the version of their layer,
+    /// so there's no reason to have the version.
+    has_obs_layer: bool,
 }
 
 pub struct InstanceShared {
@@ -400,6 +404,10 @@ pub struct CommandEncoder {
     /// If this is true, the active renderpass enabled a debug span,
     /// and needs to be disabled on renderpass close.
     rpass_debug_marker_active: bool,
+
+    /// If set, the end of the next render/compute pass will write a timestamp at
+    /// the given pool & location.
+    end_of_pass_timer_query: Option<(vk::QueryPool, u32)>,
 }
 
 impl fmt::Debug for CommandEncoder {